Add raster tiles from a local folder

Add tiled raster data pyramids from a local folder that was created with gdal2tiles.py (see https://gdal.org/en/latest/programs/gdal2tiles.html for details).

Usage

addTileFolder(
  map,
  folder,
  tms = TRUE,
  layerId = NULL,
  group = NULL,
  attribution = NULL,
  options = leaflet::tileOptions(),
  data = leaflet::getMapData(map)
)

Arguments

map

a mapview or leaflet object.

folder

the (top level) folder where the tiles (folders) reside.

tms

whether the tiles are served as TMS tiles.

layerId

the layer id.

group

the group name for the tile layer to be added to map.

attribution

the attribution text of the tile layer (HTML).

options

a list of extra options for tile layers. See tileOptions for details. When the tiles were created with gdal2tiles.py argument tms needs to be set to TRUE.

data

the data object from which the argument values are derived; by default, it is the data object provided to leaflet() initially, but can be overridden.