Update the layer controls when adding layers to an existing map.

When adding additional base layers or overlay layers to an existing map, updateLayersControl will either update the existing layers control or add a new one if map has none.

Usage

updateLayersControl(
  map,
  addBaseGroups = character(0),
  addOverlayGroups = character(0),
  position = "topleft",
  ...
)

Arguments

map

A leaflet or mapview map.

addBaseGroups

group names of base layers to be added to layers control.

addOverlayGroups

group names of overlay layers to be added to layers control.

position

position of control: "topleft", "topright", "bottomleft", or "bottomright".

...

Further arguments passed to addLayersControl.

Value

A leaflet map object.

Examples

library(leaflet)

map = leaflet() %>%
        addProviderTiles("OpenStreetMap", group = "OSM") %>%
        addProviderTiles("CartoDB.DarkMatter", group = "dark") %>%
        addCircleMarkers(data = breweries91, group = "brew")

map # no layers control

map %>%
  updateLayersControl(addBaseGroups = c("OSM", "dark"),
                      addOverlayGroups = "brew")